// packages/unified/src/remark/utils.d.ts
export declare class RemarkUtils {
    /**
     * Use this to [[Get the line offset of the frontmatter|dendron://dendron.docs/pkg.plugin-core.dev.cook#get-the-line-offset-of-the-frontmatter]]
     * Given a string representation of a Dendron note,
     * return the position of the line right after the frontmatter.
     * @param fileText file content string to traverse
     * @returns position in parsed file content right after the frontmatter
     */
    static getNodePositionPastFrontmatter(fileText: string): Position | undefined;
    static bumpHeadings(root: Parent, baseDepth: number): void;
    static findAnchors(content: string): Anchor[];
    static isHeading(node: Node, text: string, depth?: number): node is Heading;
    static isRoot(node: Node): node is Parent;
    static isParent(node: Node): node is Parent;
    static isParagraph(node: Node): node is Paragraph;
    static isTable(node: Node): node is Table;
    static isTableRow(node: Node): node is TableRow;
    static isTableCell(node: Node): node is TableCell;
    static isList(node: Node): node is List;
    static isNoteRefV2(node: Node): node is NoteRefNoteV4;
    static isImage(node: Node): node is Image;
    static isExtendedImage(node: Node): node is ExtendedImage;
    static isText(node: Node): node is Text;
    static isLink(node: Node): node is Link;
    static isWikiLink(node: Node): node is WikiLinkNoteV4;
    static isFootnoteDefinition(node: Node): node is FootnoteDefinition;
    static isFrontmatter(node: Node): node is FrontmatterContent;
    static isHTML(node: Node): node is HTML;
    static isCode(node: Node): node is Code;
    static isYAML(node: Node): node is YAML;
    static isHashTag(node: Node): node is HashTag;
    static isUserTag(node: Node): node is UserTag;
    static isNodeWithPosition<N extends Node>(node: N): node is N & {
        position: Position;
    };
    static convertLinksFromDotNotation(note: NoteProps, changes: NoteChangeEntry[]): (this: Processor) => (tree: Node, _vfile: VFile) => void;
    static convertWikiLinkToNoteUrl(note: NoteProps, changes: NoteChangeEntry[], engine: DEngineClient, dendronConfig: DendronConfig): (this: Processor) => (tree: Node, _vfile: VFile) => Promise<void>;
    static h1ToTitle(note: NoteProps, changes: NoteChangeEntry[]): (this: Processor) => (tree: Node, _vfile: VFile) => void;
    static h1ToH2(note: NoteProps, changes: NoteChangeEntry[]): (this: Processor) => (tree: Node, _vfile: VFile) => void;
    /**
     * Recursively check if two given node has identical children.
     * At each level _position_ is omitted as this can change if
     * you are comparing from two different trees.
     * @param a first {@link Node} to compare
     * @param b second {@link Node} to compare
     * @returns boolean
     */
    static hasIdenticalChildren: (a: Node, b: Node) => boolean;
    /**
     * Given a markdown AST and a target heading node,
     * Find all the node that belongs under the heading.
     * This will extract all nodes until it hits the next heading
     * with the same depth of the target heading.
     * @param tree Abstract syntax tree
     * @param startHeaderDepth Heading to target
     * @returns nodes to extract
     */
    static extractHeaderBlock(tree: Node, startHeaderDepth: number, startHeaderIndex: number, stopAtFirstHeader?: boolean): Node<import("unist").Data>[];
    /** Extract all blocks from the note which could be referenced by a block anchor.
     *
     * If those blocks already have anchors (or if they are a header), this will also find that anchor.
     *
     * @param note The note from which blocks will be extracted.
     */
    static extractBlocks({ note, config, }: {
        note: NoteProps;
        config: DendronConfig;
    }): Promise<NoteBlock[]>;
    static extractFootnoteDefs(root: Node): FootnoteDefinition[];
    /**
     * Extract frontmatter tags from note
     * @param body
     * @returns
     */
    static extractFMTags(body: string): import("yaml-unist-parser").Literal[];
    static getNoteUrl(opts: {
        config: DendronConfig;
        note: NotePropsMeta;
        vault: DVault;
        urlRoot?: string;
        anchor?: string;
    }): string;
}